+++ source/modules/interfaces/VTK/VTU_Interface.cpp
@@ -2655,7 +2655,7 @@ void VTU_Interface<dim>::WriteElementDataArrayScalarArray( const Index& key, XML
   string variableName;
   string prefix;
   ArrayVariable arrayVariable;
-  for( size_t component=0; component<key.dataDepth; component++)
+  for( uint32_t component=0u; component<key.dataDepth; component++ )
   {
       arrayTitle = "DataArray type=\"Float64\" Name=\"";
       variableName = model_.Database().Name( key );
